requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# Visualization
plotly==5.17.0

# LangChain Ecosystem
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langgraph==0.0.20
langserve==0.0.30

# Vector Database & Embeddings (Compatible versions)
chromadb==0.4.18
sentence-transformers==2.2.2
huggingface_hub==0.16.4
transformers==4.30.2
torch==2.0.1
tokenizers==0.13.3
faiss-cpu==1.7.4

# Data Processing & Analysis
pandas==2.1.3
numpy==1.24.3
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
openpyxl==3.1.2

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# HTTP Client & API
requests==2.31.0
httpx==0.25.2
aiofiles==23.1.0
aiohttp==3.8.5

# Utilities & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
typing-extensions==4.8.0
rich==13.7.0
loguru==0.7.2
python-dateutil==2.8.2
uuid==1.30
pathlib==1.0.1
psutil==5.9.6
pydantic[email]

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1

euriai